Move non-trivial definitions out of apollo-router/src/lib.rs
This is in preparation of merging apollo-router-core there
