+++ apollo-router/src/plugins/telemetry/metrics/prometheus.rs
@@ -1,8 +1,8 @@
-use crate::future::BoxFuture;
 use crate::plugins::telemetry::config::MetricsCommon;
 use crate::plugins::telemetry::metrics::{MetricsBuilder, MetricsConfigurator};
 use apollo_router_core::{http_compat, ResponseBody};
 use bytes::Bytes;
+use futures::future::BoxFuture;
 use http::StatusCode;
 use opentelemetry::sdk::Resource;
 use opentelemetry::{Key, KeyValue, Value};
